♻️ Refactor : 모듈화된 컴포넌트 병합
- 스터디 커리큘럼 불필요한 모듈화 제거
